Move moshi adapters into their own package
