+++ app/src/main/java/org/simple/clinic/util/UuidTypeConverters.kt
@@ -1,8 +1,6 @@
 package org.simple.clinic.util
 
 import androidx.room.TypeConverter
-import com.squareup.moshi.FromJson
-import com.squareup.moshi.ToJson
 import java.util.UUID
 
 class UuidRoomTypeConverter {
@@ -18,15 +16,3 @@ class UuidRoomTypeConverter {
   }
 }
 
-class UuidMoshiAdapter {
-
-  @FromJson
-  fun toUuid(value: String?): UUID? {
-    return value?.let { UUID.fromString(it) }
-  }
-
-  @ToJson
-  fun fromUuid(uuid: UUID?): String? {
-    return uuid?.toString()
-  }
-}
