+++ app/src/main/java/org/simple/clinic/util/moshi/LocalDateMoshiAdapter.kt
@@ -1,25 +1,25 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.moshi
 
-import androidx.room.TypeConverter
+import com.squareup.moshi.FromJson
+import com.squareup.moshi.ToJson
 import org.threeten.bp.LocalDate
 import org.threeten.bp.format.DateTimeFormatter
 
-class LocalDateRoomTypeConverter {
+class LocalDateMoshiAdapter {
 
   companion object {
     private val formatter = DateTimeFormatter.ISO_DATE!!
   }
 
-  @TypeConverter
+  @FromJson
   fun toLocalDate(value: String?): LocalDate? {
     return value?.let {
       return formatter.parse(value, LocalDate::from)
     }
   }
 
-  @TypeConverter
+  @ToJson
   fun fromLocalDate(date: LocalDate?): String? {
     return date?.format(formatter)
   }
 }
-
