+++ app/src/main/java/org/simple/clinic/util/moshi/OptionalMoshiAdapter.kt
@@ -1,9 +1,12 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.moshi
 
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.JsonReader
 import com.squareup.moshi.JsonWriter
 import com.squareup.moshi.Moshi
+import org.simple.clinic.util.Just
+import org.simple.clinic.util.None
+import org.simple.clinic.util.Optional
 
 import java.io.IOException
 import java.lang.reflect.ParameterizedType
